# internal/core/domain/customer.go
package domain

type Customer struct {
	ID    uint   `json:"id"`
	Name  string `json:"name" validate:"required"`
	CPF   string `json:"cpf" validate:"required"`
	Email string `json:"email" validate:"required"`
}

type CustomerResponse struct {
	Id uint `json:"id"`
}